+++ CardForm/src/main/java/com/braintreepayments/cardform/view/CardForm.java
@@ -281,8 +281,6 @@ public class CardForm extends LinearLayout implements OnCardTypeChangedListener,
         mPostalCodeIcon.setImageResource(isDarkBackground ? R.drawable.bt_ic_postal_code_dark : R.drawable.bt_ic_postal_code);
         mMobileNumberIcon.setImageResource(isDarkBackground? R.drawable.bt_ic_mobile_number_dark : R.drawable.bt_ic_mobile_number);
 
-        mExpiration.useDialogForExpirationDateEntry(activity, true);
-
         setViewVisibility(mCardholderNameIcon,  cardHolderNameVisible);
         setFieldVisibility(mCardholderName, cardHolderNameVisible);
         setViewVisibility(mCardNumberIcon, mCardNumberRequired);